+++ resources/views/components/modal/dialog.blade.php
@@ -9,7 +9,7 @@
         </div>
     </div>
 
-    <div class="px-6 py-4 bg-gray-100 text-right">
+    <div class="px-6 py-4 bg-gray-100 dark:bg-gray-800 text-right">
         {{ $footer }}
     </div>
 </x-wb::modal>