+++ resources/views/components/table.blade.php
@@ -1,0 +1,14 @@
+<div class="overflow-x-auto">
+    <div class="align-middle inline-block min-w-full">
+        <div class="shadow overflow-hidden border-b border-gray-200 dark:border-gray-500 sm:rounded-b-sm">
+            <table class="min-w-full divide-y divide-gray-200 dark:divide-gray-500">
+                <thead class="bg-gray-100 dark:bg-slate-700 select-none">
+                    {{ $head }}
+                </thead>
+                <tbody class="bg-white dark:bg-slate-600 divide-y divide-gray-200 dark:divide-gray-600">
+                    {{ $body }}
+                </tbody>
+            </table>
+        </div>
+    </div>
+</div>