+++ resources/views/components/table/row.blade.php
@@ -1,0 +1,3 @@
+<tr class="text-xs">
+    {{ $slot }}
+</tr>